requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
google-genai
fastmcp